kali_execution_server/requirements_server.txt
fastapi
uvicorn[standard]
docker
paramiko
kali-driver
openai
python-dotenv